Restructure DddCore to have namespace like Entities, Persistence, Validation, Identity, Cqrs, and DomainEvents
