+++ account-management/src/Domain/Tenants/TenantValidation.cs
@@ -1,4 +1,4 @@
-using PlatformPlatform.Foundation.DddCore;
+using PlatformPlatform.Foundation.DddCore.Validation;
 
 namespace PlatformPlatform.AccountManagement.Domain.Tenants;
