+++ account-management/src/Foundation/DependencyInjection.cs
@@ -3,9 +3,8 @@ using MediatR;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using PlatformPlatform.Foundation.Application.DomainEvents;
-using PlatformPlatform.Foundation.Application.Persistence;
-using PlatformPlatform.Foundation.DddCore;
+using PlatformPlatform.Foundation.DddCore.DomainEvents;
+using PlatformPlatform.Foundation.DddCore.Persistence;
 using PlatformPlatform.Foundation.Infrastructure;
 
 namespace PlatformPlatform.Foundation;
