+++ account-management/src/Foundation/WebApi/CommandResultExtensions.cs
@@ -1,5 +1,5 @@
 using Microsoft.AspNetCore.Http;
-using PlatformPlatform.Foundation.Application;
+using PlatformPlatform.Foundation.DddCore.Cqrs;
 
 namespace PlatformPlatform.Foundation.WebApi;
