+++ account-management/tests/Foundation.Tests/Application/Persistence/UnitOfWorkPipelineBehaviorTests.cs
@@ -1,7 +1,6 @@
 using MediatR;
 using NSubstitute;
-using PlatformPlatform.Foundation.Application.Persistence;
-using PlatformPlatform.Foundation.DddCore;
+using PlatformPlatform.Foundation.DddCore.Persistence;
 using PlatformPlatform.Foundation.Tests.TestEntities;
 using Xunit;
