+++ account-management/tests/Foundation.Tests/Domain/EntityTests.cs
@@ -1,6 +1,7 @@
 using FluentAssertions;
 using JetBrains.Annotations;
-using PlatformPlatform.Foundation.DddCore;
+using PlatformPlatform.Foundation.DddCore.Entities;
+using PlatformPlatform.Foundation.DddCore.Identity;
 using Xunit;
 
 namespace PlatformPlatform.Foundation.Tests.Domain;
