+++ account-management/tests/Foundation.Tests/Infrastructure/RepositoryTests.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
-using PlatformPlatform.Foundation.DddCore;
+using PlatformPlatform.Foundation.DddCore.Identity;
 using PlatformPlatform.Foundation.Tests.TestEntities;
 using Xunit;
