+++ account-management/tests/Foundation.Tests/Infrastructure/UnitOfWorkTests.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
-using PlatformPlatform.Foundation.DddCore;
+using PlatformPlatform.Foundation.DddCore.Persistence;
 using PlatformPlatform.Foundation.Infrastructure;
 using PlatformPlatform.Foundation.Tests.TestEntities;
 using Xunit;
