+++ account-management/tests/Tests/ArchitectureTests/PublicClassesTests.cs
@@ -3,7 +3,7 @@ using NetArchTest.Rules;
 using PlatformPlatform.AccountManagement.Application;
 using PlatformPlatform.AccountManagement.Domain;
 using PlatformPlatform.AccountManagement.Infrastructure;
-using PlatformPlatform.Foundation.Application;
+using PlatformPlatform.Foundation.DddCore.Cqrs;
 using Xunit;
 
 namespace PlatformPlatform.AccountManagement.Tests.ArchitectureTests;
