+++ account-management/tests/Tests/WebApi/Endpoints/TenantEndpointsTests.cs
@@ -10,7 +10,7 @@ using PlatformPlatform.AccountManagement.Domain.Tenants;
 using PlatformPlatform.AccountManagement.Infrastructure;
 using PlatformPlatform.AccountManagement.Tests.Infrastructure;
 using PlatformPlatform.AccountManagement.WebApi.Endpoints;
-using PlatformPlatform.Foundation.DddCore;
+using PlatformPlatform.Foundation.DddCore.Validation;
 using Xunit;
 
 namespace PlatformPlatform.AccountManagement.Tests.WebApi.Endpoints;
