+++ backend-foundation/src/DddCore/Identity/StronglyTypedId.cs
@@ -1,32 +1,4 @@
-using System.Reflection;
-
-namespace PlatformPlatform.Foundation.DddCore;
-
-/// <summary>
-///     This is a special version of <see cref="StronglyTypedId{T}" /> for longs which is the recommended type to use.
-///     It uses the <see cref="IdGenerator" />  to create IDs that are chronological and guaranteed to be unique.
-/// </summary>
-public abstract record StronglyTypedId<T>(long Value) : StronglyTypedId<long, T>(Value)
-    where T : StronglyTypedId<T>
-{
-    public static T NewId()
-    {
-        var newValue = IdGenerator.NewId();
-        return FormLong(newValue);
-    }
-
-    public static T FromString(string value)
-    {
-        var newValue = Convert.ToInt64(value);
-        return FormLong(newValue);
-    }
-
-    private static T FormLong(long newValue)
-    {
-        return (T) Activator.CreateInstance(typeof(T),
-            BindingFlags.Instance | BindingFlags.Public, null, new object[] {newValue}, null)!;
-    }
-}
+namespace PlatformPlatform.Foundation.DddCore.Identity;
 
 /// <summary>
 ///     StronglyTypedId is an abstract record type for creating strongly typed IDs with a specified value type. It makes
